clean up read & decode story
